refactor: isolate retryable client creation
